오브젝트 컨트롤 추가 상속구조 작성, 타겟 공격 개선
오브젝트 컨트롤 추가
유닛컨트롤 오브젝트 컨트롤 자식 객체로 설정
오브젝트 컨트롤을 상속하는 빌딩 컨트롤 추가
빌딩 매니저 추가
빌딩 정보를 저장하는 빌딩인포XML 추가
빌딩정보XML을 위한 코드 작성
빌딩프리팹 생성

+++ BattleProject/Assets/Resources/Script/TestScript/UnitInfo.cs
@@ -182,7 +182,7 @@ namespace GameSys
                 }
             }
 
-            private Dictionary<int, JobInfo> jobDic = new Dictionary<int, JobInfo>();
+            private Dictionary<byte, JobInfo> jobDic = new Dictionary<byte, JobInfo>();
 
             public void AddJob(JobInfo info)
             {
@@ -191,7 +191,7 @@ namespace GameSys
                 jobDic.Add(info.ID, info);
             }
 
-            public JobInfo Job(int ID)
+            public JobInfo Job(byte ID)
             {
                 return jobDic[ID];
             }
@@ -216,7 +216,7 @@ namespace GameSys
                 }
             }
 
-            private Dictionary<int, JobEquipInfo> jobEquipDic = new Dictionary<int, JobEquipInfo>();
+            private Dictionary<byte, JobEquipInfo> jobEquipDic = new Dictionary<byte, JobEquipInfo>();
 
             public void AddJobEquip(JobEquipInfo info)
             {
@@ -225,7 +225,7 @@ namespace GameSys
                 jobEquipDic.Add(info.ID, info);
             }
 
-            public JobEquipInfo JobEquip(int ID)
+            public JobEquipInfo JobEquip(byte ID)
             {
                 return jobEquipDic[ID];
             }
